Basic template and all the modules setup.
This was the first real commit where I added all the required modules and setup the file directories.

I'm going to use express for this project and maybe mongodb. Don't know if I'm going to bother fixing with databases, or I should just use json file. (I'm not going to store passwords so it might be fine for the item prices)

This projects frontend framework will be bootsrap 4 since it's easy to use and I won't go into react for this (I don't know react xD).

The basics with the server should be working, you can connect to the website and it will render correctly.

Tomorrow I will so it will display all the items instead of the single sample item that there is now.

Other:

Looking forward to making this website and contributing to the Civex community.