Readme, search function, itemlist, offer examples
Updated the readme file and added some basic information about the project. I also added a coupple of screenshots if someone is interested.

Added searchfunction on the itemlist page.

Added all the items as data and made them export into the client. All the images should be fairly easy to render since they are very small.

Added an example page when someone adds a trade when looking for a trade partner.

Added selling.ejs

Added so it's connecting to a database.

Minor fixes.