Frontend Hell
I sat down hours trying to figure out the frontend design for everything.

I also made the page with the input forms for adding a selling listing to the database. I'm soon going to connect everything to a database and will work with login forms..

I will also have to write a shit ton of text for user agreement and everything.